requirements.txt
flask>=3.0
flask-cors>=4.0
PyJWT>=2.8
bcrypt>=4.1
scikit-learn>=1.4
numpy>=1.26
Pillow>=10.0
